[Rename] 'BookBlock' 컴포넌트와 'BookBlockList' 컴포넌트 추상 계층 변경 외 작업 진행
